Update MetaWear hook
