+++ hooks/index.ts
@@ -1,3 +1,4 @@
 export * from "./useAuth";
 export * from "./useColorScheme";
+export * from "./useMetaWear";
 export * from "./useThemeColor";
